requirements.txt
pynput
cherrypy
python-dotenv
